pyproject.toml
[project]
name = "anemometer-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "djoser>=2.3.1",
    "drf-nested-routers>=0.94.2",
    "drf-spectacular>=0.28.0",
]
